Implement AWS Transcribe Medical Direct API with WebSocket and HTTP/2
